requirements.txt
lxml
loguru
click
toml
pillow
PyMuPDF
pandas
gooey
pytesseract
natsort